remove redundant component foundations
