+++ client/components/EDA.jsx
@@ -1,7 +1,44 @@
 import React from 'react'
-import Foundations from './Foundations'
 
 class EDA extends React.Component {
+    state = {
+        id: '',
+    }
+
+    handleClick = (e) => {
+        const id = e.target.getAttribute('id')
+        if (this.state.id == id) {
+            this.setState ({
+                id: '',
+            })
+        } else {
+            this.setState ({
+                id: id,
+            })
+        }
+    }
+
+    renderContent = (id) => {
+        switch (id){
+            case "1":
+                return <Blog/>
+            case "2":
+                return <Minesweeper/>
+            default:
+                return (
+                    <div>
+                        <h3>Foundations Contents</h3>
+                        <p>
+                            Blurb about foundations
+                        </p>
+                        <p>
+                            make pages flash clickable image when hover??
+                        </p>
+                    </div>
+                )
+        } 
+    }
+
     render() {
         return (
             <>
@@ -13,7 +50,23 @@ class EDA extends React.Component {
                     HTML, CSS, JavaScript and Git was covered in preperations for 
                     the intensive Bootcamp that was to come. 
                 </p>
-                <Foundations/>
+            <div className="container">
+                <div className="list">
+                    <div className="text">
+                        <p className="hover">
+                            <a id="1" onClick={this.handleClick}>Blog</a>
+                        </p>
+                        <p className="hover">
+                            <a id="2" onClick={this.handleClick}>Minesweeper</a>
+                        </p>
+                    </div>
+                </div>
+                <div className="content">
+                    <div className="text">
+                        {this.renderContent(this.state.id)}
+                    </div>
+                </div>
+            </div>
             </>
         )
     }
