+++ server/public/main.css
@@ -1,7 +1,10 @@
 body {
     font-family: 'Source Sans Pro', sans-serif;
     background-color: rgb(191, 236, 250);
-    margin: 5% 15%;
+    margin: 0 auto;
+    width: 100%;
+    max-width: 700px;
+    height: auto;
 }
 
 .navbar {
